Refactor: rename robohash to cathash
