+++ apps/desktop/desktop/src/main/java/bisq/desktop/main/content/chat/BaseChatController.java
@@ -31,7 +31,7 @@ import bisq.desktop.common.threading.UIThread;
 import bisq.desktop.common.view.Navigation;
 import bisq.desktop.common.view.NavigationController;
 import bisq.desktop.components.controls.BisqIconButton;
-import bisq.desktop.components.robohash.RoboHash;
+import bisq.desktop.components.cathash.CatHash;
 import bisq.desktop.main.content.chat.sidebar.ChannelSidebar;
 import bisq.desktop.main.content.chat.sidebar.UserProfileSidebar;
 import bisq.desktop.main.content.components.chatMessages.ChatMessagesComponent;
@@ -147,7 +147,7 @@ public abstract class BaseChatController<V extends BaseChatView, M extends BaseC
     protected void applyPeersIcon(PrivateChatChannel<?> privateChatChannel) {
         if (privateChatChannel instanceof TwoPartyPrivateChatChannel) {
             TwoPartyPrivateChatChannel twoPartyPrivateChatChannel = (TwoPartyPrivateChatChannel) privateChatChannel;
-            Image image = RoboHash.getImage(twoPartyPrivateChatChannel.getPeer().getPubKeyHash());
+            Image image = CatHash.getImage(twoPartyPrivateChatChannel.getPeer().getPubKeyHash());
             ImageView imageView = new ImageView(image);
             imageView.setFitWidth(35);
             imageView.setFitHeight(35);
@@ -167,13 +167,13 @@ public abstract class BaseChatController<V extends BaseChatView, M extends BaseC
                     left = bisqEasyOpenTradeChannel.getPeer();
                     right = bisqEasyOpenTradeChannel.getMediator().get();
                 }
-                ImageView leftImageView = new ImageView(RoboHash.getImage(left.getPubKeyHash()));
+                ImageView leftImageView = new ImageView(CatHash.getImage(left.getPubKeyHash()));
                 leftImageView.setFitWidth(35);
                 leftImageView.setFitHeight(35);
                 Button leftIconButton = BisqIconButton.createIconButton(leftImageView);
                 leftIconButton.setMouseTransparent(true);
 
-                ImageView rightImageView = new ImageView(RoboHash.getImage(right.getPubKeyHash()));
+                ImageView rightImageView = new ImageView(CatHash.getImage(right.getPubKeyHash()));
                 rightImageView.setFitWidth(35);
                 rightImageView.setFitHeight(35);
                 Button rightIconButton = BisqIconButton.createIconButton(rightImageView);
@@ -184,7 +184,7 @@ public abstract class BaseChatController<V extends BaseChatView, M extends BaseC
                 hBox.setAlignment(Pos.CENTER_LEFT);
                 model.getChannelIconNode().set(hBox);
             } else {
-                Image image = RoboHash.getImage(bisqEasyOpenTradeChannel.getPeer().getPubKeyHash());
+                Image image = CatHash.getImage(bisqEasyOpenTradeChannel.getPeer().getPubKeyHash());
                 ImageView imageView = new ImageView(image);
                 imageView.setFitWidth(35);
                 imageView.setFitHeight(35);
